requirements.txt
arrow==1.2.3
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.54.1
fsspec==2024.10.0
gunicorn==23.0.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0.2
requests==2.28.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SpeechRecognition==3.9.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
ultralytics-thop==2.0.10
urllib3==1.26.14
Werkzeug==3.1.3
